requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.0
bson==0.5.10
Pillow==10.1.0
google-generativeai==0.7.2
python-multipart==0.0.6
